Replace free variables in expressions by functions directly after parsing.
Previously, an `EVar` in the AST meant either a monomorphic function
invocation or a local variable. This change removes this ambiguity.
As a consequence, the type checkers were simplified.

The pretty printers now mark functions with a type instantiation
list even if it is empty. This helps to tell variables and functions
apart.